requirements.txt
apilogs==1.2
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.1.0
awscli==1.16.272
awslogs==0.11.0
beautifulsoup4==4.4.0
boto3==1.10.2
botocore==1.13.8
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
cheroot==8.2.1
Click==7.0
colorama==0.3.9
cryptography==2.8
cycler==0.10.0
docutils==0.14
flaky==3.6.1
Flask==1.1.1
Flask-Cors==3.0.8
git-python==1.0.3
gitdb2==2.0.6
GitPython==3.0.5
idna==2.8
importlib-metadata==0.19
imutils==0.5.3
isort==4.3.21
itsdangerous==1.1.0
jaraco.functools==3.0.0
Jinja2==2.11.1
jmespath==0.9.4
jwt==0.6.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
lib==3.0.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
more-itertools==8.1.0
numpy==1.18.1
opencv-contrib-python==4.1.2.30
opencv-python==4.1.2.30
packaging==20.1
pathlib==1.0.1
Pillow==7.0.0
pipfile==0.0.2
pluggy==0.13.1
polling2==0.4.5
py==1.8.1
pyasn1==0.4.6
pycparser==2.19
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pytest==5.3.5
pytest-html==2.0.1
pytest-metadata==1.8.0
python-dateutil==2.8.0
PyYAML==5.1.2
requests==2.22.0
rsa==3.4.2
s3transfer==0.2.1
six==1.12.0
smmap2==2.0.5
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.3
wcwidth==0.1.8
web.py==0.40
Werkzeug==1.0.0
wrapt==1.11.2
zipp==0.5.2
